paraswap: update task with deployer v3
