--- tasks/2023011300-paraswap-fee-redistributor/output/polygon.json
@@ -1,7 +1,0 @@
-{
-  "SmartVaultDeployer": "0x91000AD2279BFe320d12721c5c0903b1FB09960D",
-  "ERC20Claimer": "0xD68C6Ed978D4c115AB25E13879440366d17a71BA",
-  "NativeClaimer": "0x05Cd5EcF62C18058C11F44f4DC8caD867aAEe705",
-  "SwapFeeSetter": "0x5853dE8e01A0320b4C274f804AcB1DE5B40cDF8d",
-  "SmartVault": "0x00f076bDb53724fEd5D0338c7fb76AAD3152fc58"
-}
