--- tasks/2023011300-paraswap-fee-redistributor/test/.config.json
@@ -1,20 +1,0 @@
-{
-  "fork": {
-    "arbitrum": 0,
-    "avalanche": 0,
-    "bsc": 0,
-    "fantom": 0,
-    "mainnet": 0,
-    "optimism": 0,
-    "polygon": 38399232
-  },
-  "deployed": {
-    "arbitrum": 0,
-    "avalanche": 0,
-    "bsc": 0,
-    "fantom": 0,
-    "mainnet": 0,
-    "optimism": 0,
-    "polygon": 0
-  }
-}
