+++ tasks/2023020101-paraswap-fee-redistributor/test/.config.json
@@ -1,0 +1,12 @@
+{
+  "fork": {
+    "arbitrum": 0,
+    "avalanche": 0,
+    "bsc": 0,
+    "fantom": 0,
+    "mainnet": 0,
+    "optimism": 0,
+    "polygon": 0
+  },
+  "deployed": {}
+}
